pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "databricks-pipeline-template"
version = "0.1.0"
description = "Databricks medallion pipeline template"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.20.0",
    "delta-spark>=3.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",      # Retry logic for API calls
    "structlog>=24.0.0",    # Structured logging
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "responses>=0.25.0",    # HTTP mocking for tests
    "freezegun>=1.5.0",     # Time mocking for tests
]

[tool.setuptools.packages.find]
include = ["ingestion*", "pipeline*", "scripts*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501", "N812"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["ingestion/tests", "pipeline/tests"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["ingestion/src", "pipeline"]
omit = ["*/tests/*"]
